fix: compare `UrlOrPath` (#618)
This is a more general fix for
https://github.com/prefix-dev/pixi/pull/1224 .

The basic idea is that you can "canonicalize" a `UrlOrPath`. If the
stored Url represents an absolute path, it is converted into a path
instead. This is not completely trivial because the `Url::to_file_path`
method behaves differently depending on the architecture. I had to
reimplement this method.

When checking the equality of two `UrlOrPath`s they are first
canonicalized so that comparing a `file:///...` and a `/..` url and path
works properly.

`UrlOrPath`s are not canonicalized before serialization, this ensures
that loading a lock file and saving it again will result in the same
lock file. However, when _creating_ a lock file I think it makes sense
to canonicalize any instance you create.